+++ internal/codeintel/uploads/internal/store/store_commits_test.go
@@ -270,6 +270,15 @@ func insertUploads(t testing.TB, db database.DB, uploads ...shared.Upload) {
 	}
 }
 
+func deleteUploads(t testing.TB, db database.DB, uploads ...int) {
+	for _, upload := range uploads {
+		query := sqlf.Sprintf(`DELETE FROM lsif_uploads WHERE id = %s`, upload)
+		if _, err := db.ExecContext(context.Background(), query.Query(sqlf.PostgresBindVar), query.Args()...); err != nil {
+			t.Fatalf("unexpected error while deleting upload: %s", err)
+		}
+	}
+}
+
 // insertRepo creates a repository record with the given id and name. If there is already a repository
 // with the given identifier, nothing happens
 func insertRepo(t testing.TB, db database.DB, id int, name string) {
@@ -309,6 +318,36 @@ func getUploadStates(db database.DB, ids ...int) (map[int]string, error) {
 	)
 
 	return scanStates(db.QueryContext(context.Background(), q.Query(sqlf.PostgresBindVar), q.Args()...))
+}
+
+// insertVisibleAtTip populates rows of the lsif_uploads_visible_at_tip table for the given repository
+// with the given identifiers. Each upload is assumed to refer to the tip of the default branch. To mark
+// an upload as protected (visible to _some_ branch) butn ot visible from the default branch, use the
+// insertVisibleAtTipNonDefaultBranch method instead.
+func insertVisibleAtTip(t testing.TB, db database.DB, repositoryID int, uploadIDs ...int) {
+	insertVisibleAtTipInternal(t, db, repositoryID, true, uploadIDs...)
+}
+
+// insertVisibleAtTipNonDefaultBranch populates rows of the lsif_uploads_visible_at_tip table for the
+// given repository with the given identifiers. Each upload is assumed to refer to the tip of a branch
+// distinct from the default branch or a tag.
+func insertVisibleAtTipNonDefaultBranch(t testing.TB, db database.DB, repositoryID int, uploadIDs ...int) {
+	insertVisibleAtTipInternal(t, db, repositoryID, false, uploadIDs...)
+}
+
+func insertVisibleAtTipInternal(t testing.TB, db database.DB, repositoryID int, isDefaultBranch bool, uploadIDs ...int) {
+	var rows []*sqlf.Query
+	for _, uploadID := range uploadIDs {
+		rows = append(rows, sqlf.Sprintf("(%s, %s, %s)", repositoryID, uploadID, isDefaultBranch))
+	}
+
+	query := sqlf.Sprintf(
+		`INSERT INTO lsif_uploads_visible_at_tip (repository_id, upload_id, is_default_branch) VALUES %s`,
+		sqlf.Join(rows, ","),
+	)
+	if _, err := db.ExecContext(context.Background(), query.Query(sqlf.PostgresBindVar), query.Args()...); err != nil {
+		t.Fatalf("unexpected error while updating uploads visible at tip: %s", err)
+	}
 }
 
 // scanStates scans pairs of id/states from the return value of `*Store.query`.